(refactor): restructure project as lib with CLI binary
I moved all modules from src/socks5 to src/ root. Then, created a
library interface in src/lib.rs with public API. Moved main.rs CLI
application to src/bin/miniproxy.rs. Removed all TCP proxy
functionality. Updated all internal imports to reflect new structure.
Updated cargo.toml to match changes with lib and bin configuration.
