+++ Cargo.toml
@@ -2,7 +2,15 @@
 name = "miniproxy"
 version = "0.1.0"
 edition = "2024"
-description = "A lightweight TCP and SOCKS5 proxy"
+description = "A lightweight SOCKS5 proxy"
+
+[lib]
+name = "miniproxy"
+path = "src/lib.rs"
+
+[[bin]]
+name = "miniproxy"
+path = "src/bin/miniproxy.rs"
 
 [dependencies]
 clap = { version = "4.5.39", features = ["derive"] }
